@@ -11,17 +11,17 @@
 
     <properties>
         <!-- Compiler and Source settings -->
-        <java.version>11</java.version>
-        <maven.compiler.target>${java.version}</maven.compiler.target>
-        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <java.version>17</java.version>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <!-- Library versions -->
-        <spring.version>2.4.5</spring.version>
-        <security.version>5.5.0</security.version>
-        <thymeleaf.version>3.0.4.RELEASE</thymeleaf.version>
+        <spring.version>3.0.4</spring.version>
+        <security.version>6.0.2</security.version>
+        <thymeleaf.version>3.1.1.RELEASE</thymeleaf.version>
         <lombok.version>1.18.24</lombok.version>
-        <flyway.version>8.1.0</flyway.version>
+        <flyway.version>9.16.1</flyway.version>
         <databind.version>2.14.0</databind.version>
         <bouncycastle.version>1.64</bouncycastle.version>
         <codec.version>1.15</codec.version>
